requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Google Drive integration
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
chardet==5.2.0

# AI/ML dependencies
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
openai==1.35.7
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1

# Utility dependencies
pandas==2.2.2
markdown==3.6
jinja2==3.1.4
python-multipart==0.0.9

# PDF generation
pdfkit==1.0.0

# Scheduling (optional for cron alternative)
schedule==1.2.2

# Development dependencies
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.1.0
mypy==1.10.1
